+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 - ``dtool.clickutils.info_from_path`` function
 - ``dtool.DataSet.item_path_from_hash`` method
 - ``dtool.filehasher.md5sum`` function
+- ``dtool.Manifest`` class
 
 
 Changed